Resolved (#649)
* Prune failed test lists
* Minor scripts updates

+++ scripts/unit_tests.sh
@@ -1,31 +1,31 @@
 #!/bin/bash
 
-# set -e
+set -e
 
 # Check if at least one argument is provided
 if [ "$#" -lt 1 ]; then
-    echo "Usage: $0 <debug|release> <llvm-15|llvm-16|llvm-17> ..."
-    exit 1
+  echo "Usage: $0 <debug|release> <llvm-15|llvm-16|llvm-17> [skip-build]"
+  exit 1
 fi
 
 # Check if the first argument is either "debug" or "release"
 if [ "$1" != "debug" ] && [ "$1" != "release" ]; then
-    echo "Error: Invalid argument. Must be either 'debug' or 'release'."
-    exit 1
+  echo "Error: Invalid argument. Must be either 'debug' or 'release'."
+  exit 1
 fi
 
 # Set the build type based on the argument
 build_type=$(echo "$1" | tr '[:lower:]' '[:upper:]')
 
 if [ "$2" == "llvm-15" ]; then
-    LLVM=llvm-15
-    CLANG=clang/clang15-spirv-omp
+  LLVM=llvm-15
+  CLANG=clang/clang15-spirv-omp
 elif [ "$2" == "llvm-16" ]; then
-    LLVM=llvm-16
-    CLANG=clang/clang16-spirv-omp
+  LLVM=llvm-16
+  CLANG=clang/clang16-spirv-omp
 elif [ "$2" == "llvm-17" ]; then
-    LLVM=llvm-17
-    CLANG=clang/clang17-spirv-omp
+  LLVM=llvm-17
+  CLANG=clang/clang17-spirv-omp
 else
   echo "$2"
   echo "Invalid 2nd argument. Use either 'llvm-15', 'llvm-16' or 'llvm-17'."
@@ -33,16 +33,16 @@ else
 fi
 
 source /etc/profile.d/modules.sh &> /dev/null
-source /opt/intel/oneapi/setvars.sh &> /dev/null
 export MODULEPATH=$MODULEPATH:/home/pvelesko/modulefiles:/opt/intel/oneapi/modulefiles:/opt/modulefiles
 export IGC_EnableDPEmulation=1
 export OverrideDefaultFP64Settings=1
 export CHIP_LOGLEVEL=err
 export POCL_KERNEL_CACHE=0
+export CHIP_L0_COLLECT_EVENTS_TIMEOUT=30
 
 # Use OpenCL for building/test discovery to prevent Level Zero from being used in multi-thread/multi-process environment
 module load $CLANG
-module load intel/igc intel/neo
+module load intel/compute-runtime intel/opencl
 module load opencl/pocl
 output=$(clinfo -l 2>&1 | grep "Platform #0")
 echo $output
@@ -57,35 +57,42 @@ if [ -z "$output" ]; then
     exit 1
 fi
 
-rm -rf HIPCC
-rm -rf HIP
-rm -rf bitcode/ROCm-Device-Libs
-rm -rf hip-tests
-rm -rf hip-testsuite
+if [ "$3" == "skip-build" ]; then
+  echo "Skipping build step"
+  cd build
+else
+  # Build the project
+  echo "Building project..."
+  rm -rf HIPCC
+  rm -rf HIP
+  rm -rf bitcode/ROCm-Device-Libs
+  rm -rf hip-tests
+  rm -rf hip-testsuite
 
-git submodule update --init
-rm -rf build
-rm -rf *_result.txt
-mkdir build
-cd build
+  git submodule update --init
+  rm -rf build
+  rm -rf *_result.txt
+  mkdir build
+  cd build
 
-echo "building with $CLANG"
-cmake ../ -DCMAKE_BUILD_TYPE="$build_type" &> /dev/null
-make all build_tests install -j 20 #&> /dev/null
-echo "chipStar build complete." 
+  echo "building with $CLANG"
+  cmake ../ -DCMAKE_BUILD_TYPE="$build_type" &> /dev/null
+  make all build_tests install -j 20 #&> /dev/null
+  echo "chipStar build complete." 
 
+  # # Build libCEED
+  # export CHIPSTAR_INSTALL_DIR=`pwd`/install # set CHIPSTAR_INSTALL_DIR to current build dir
+  # export LIBCEED_DIR=`pwd`/libCEED
+  # ../scripts/compile_libceed.sh ${CHIPSTAR_INSTALL_DIR}
+fi
 
-# Build libCEEDA
-export CHIPSTAR_INSTALL_DIR=`pwd`/install # set CHIPSTAR_INSTALL_DIR to current build dir
-export LIBCEED_DIR=`pwd`/libCEED
-../scripts/compile_libceed.sh ${CHIPSTAR_INSTALL_DIR}
 module unload opencl/pocl
 # module load HIP/hipBLAS/main/release # for libCEED NOTE: Must be after build step otherwise it will cause link issues.
 # # Test PoCL CPU
 # echo "begin cpu_pocl_failed_tests"
 # module load opencl/pocl
 # module list
-# ctest --timeout 600 -j 20 --output-on-failure -E "`cat ./test_lists/cpu_pocl_failed_tests.txt`" | tee cpu_pocl_make_check_result.txt
+# ctest --timeout 2000 -j 20 --output-on-failure -E "`cat ./test_lists/cpu_pocl_failed_tests.txt`" | tee cpu_pocl_make_check_result.txt
 # module unload opencl/pocl
 # echo "end cpu_pocl_failed_tests"
 
@@ -93,7 +100,7 @@ module unload opencl/pocl
 echo "begin igpu_level0_failed_reg_tests"
 module load level-zero/igpu
 module list
-CHIP_L0_IMM_CMD_LISTS=OFF ctest --timeout 600 -j 20--output-on-failure -E "`cat ./test_lists/igpu_level0_failed_reg_tests.txt`" | tee igpu_level0_reg_make_check_result.txt
+CHIP_L0_IMM_CMD_LISTS=OFF ctest --timeout 2000 -j 20--output-on-failure -E "`cat ./test_lists/igpu_level0_failed_reg_tests.txt`" | tee igpu_level0_reg_make_check_result.txt
 # pushd ${LIBCEED_DIR}
 # make FC= CC=clang CXX=clang++ BACKENDS="/gpu/hip/ref /gpu/hip/shared /gpu/hip/gen" prove -j 20 PROVE_OPS="-j" | tee dgpu_level0_reg_make_check_result.txt
 # popd
@@ -105,7 +112,7 @@ echo "end igpu_level0_failed_reg_tests"
 # echo "begin igpu_level0_failed_imm_tests"
 # module load level-zero/igpu
 # module list
-# CHIP_L0_IMM_CMD_LISTS=ON ctest --timeout 600 -j 20--output-on-failure -E "`cat ./test_lists/igpu_level0_failed_imm_tests.txt`" | tee igpu_level0_imm_make_check_result.txt
+# CHIP_L0_IMM_CMD_LISTS=ON ctest --timeout 2000 -j 20--output-on-failure -E "`cat ./test_lists/igpu_level0_failed_imm_tests.txt`" | tee igpu_level0_imm_make_check_result.txt
 # # pushd ${LIBCEED_DIR}
 # # make FC= CC=clang CXX=clang++ BACKENDS="/gpu/hip/ref /gpu/hip/shared /gpu/hip/gen" prove -j 20 PROVE_OPS="-j" | tee dgpu_level0_imm_make_check_result.txt
 # # popd
@@ -116,7 +123,7 @@ echo "end igpu_level0_failed_reg_tests"
 echo "begin dgpu_level0_failed_reg_tests"
 module load level-zero/dgpu
 module list
-CHIP_L0_IMM_CMD_LISTS=OFF ctest --timeout 600 -j 20 --output-on-failure -E "`cat ./test_lists/dgpu_level0_failed_reg_tests.txt`" | tee dgpu_level0_reg_make_check_result.txt
+CHIP_L0_IMM_CMD_LISTS=OFF ctest --timeout 2000 -j 20 --output-on-failure -E "`cat ./test_lists/dgpu_level0_failed_reg_tests.txt`" | tee dgpu_level0_reg_make_check_result.txt
 # pushd ${LIBCEED_DIR}
 # HIP_DIR=${CHIPSTAR_INSTALL_DIR} make FC= CC=clang CXX=clang++ BACKENDS="/gpu/hip/ref /gpu/hip/shared /gpu/hip/gen" prove -j 20 PROVE_OPS="-j" | tee dgpu_level0_reg_make_check_result.txt
 # popd
@@ -127,7 +134,7 @@ echo "end dgpu_level0_failed_reg_tests"
 echo "begin dgpu_level0_failed_imm_tests"
 module load level-zero/dgpu
 module list
-CHIP_L0_IMM_CMD_LISTS=ON ctest --timeout 600 -j 20 --output-on-failure -E "`cat ./test_lists/dgpu_level0_failed_imm_tests.txt`" | tee dgpu_level0_imm_make_check_result.txt
+CHIP_L0_IMM_CMD_LISTS=ON ctest --timeout 2000 -j 20 --output-on-failure -E "`cat ./test_lists/dgpu_level0_failed_imm_tests.txt`" | tee dgpu_level0_imm_make_check_result.txt
 # pushd ${LIBCEED_DIR}
 # HIP_DIR=${CHIPSTAR_INSTALL_DIR} make FC= CC=clang CXX=clang++ BACKENDS="/gpu/hip/ref /gpu/hip/shared /gpu/hip/gen" prove -j 20 PROVE_OPS="-j" | tee dgpu_level0_imm_make_check_result.txt
 # popd
@@ -136,13 +143,13 @@ echo "end dgpu_level0_failed_imm_tests"
 
 # Test OpenCL iGPU
 echo "begin igpu_opencl_failed_tests"
-module load opencl/intel-igpu
+module load opencl/igpu
 module list
-ctest --timeout 600 -j 20 --output-on-failure -E "`cat ./test_lists/igpu_opencl_failed_tests.txt`" | tee igpu_opencl_make_check_result.txt
+ctest --timeout 2000 -j 20 --output-on-failure -E "`cat ./test_lists/igpu_opencl_failed_tests.txt`" | tee igpu_opencl_make_check_result.txt
 #pushd ${LIBCEED_DIR}
 #make FC= CC=clang CXX=clang++ BACKENDS="/gpu/hip/ref /gpu/hip/shared /gpu/hip/gen" prove -j 20 PROVE_OPS="-j" | tee igpu_opencl_make_check_result.txt
 #popd
-module unload opencl/intel-igpu
+module unload opencl/igpu
 echo "end igpu_opencl_failed_tests"
 
 # Test OpenCL dGPU
@@ -150,7 +157,7 @@ echo "begin dgpu_opencl_failed_tests"
 module load intel/opencl # sets ICD
 module load opencl/dgpu # sets CHIP_BE
 module list
-ctest --timeout 600 -j 20 --output-on-failure -E "`cat ./test_lists/dgpu_opencl_failed_tests.txt`" | tee dgpu_opencl_make_check_result.txt
+ctest --timeout 2000 -j 20 --output-on-failure -E "`cat ./test_lists/dgpu_opencl_failed_tests.txt`" | tee dgpu_opencl_make_check_result.txt
 # pushd ${LIBCEED_DIR}
 # HIP_DIR=${CHIPSTAR_INSTALL_DIR} make FC= CC=clang CXX=clang++ BACKENDS="/gpu/hip/ref /gpu/hip/shared /gpu/hip/gen" prove -j 20 PROVE_OPS="-j" | tee dgpu_opencl_make_check_result.txt
 # popd
@@ -161,7 +168,7 @@ echo "end dgpu_opencl_failed_tests"
 # echo "begin cpu_opencl_failed_tests"
 # module load opencl/cpu
 # module list
-# ctest --timeout 600 -j 20 --output-on-failure -E "`cat ./test_lists/cpu_opencl_failed_tests.txt`" | tee cpu_opencl_make_check_result.txt
+# ctest --timeout 2000 -j 20 --output-on-failure -E "`cat ./test_lists/cpu_opencl_failed_tests.txt`" | tee cpu_opencl_make_check_result.txt
 # module unload opencl/cpu
 # echo "end cpu_opencl_failed_tests"
 
@@ -201,7 +208,7 @@ for test_result in dgpu_opencl_make_check_result.txt \
                    dgpu_level0_imm_make_check_result.txt
 do
   echo -n "${test_result}: "
-  check_tests "${test_result}"
+  check_tests "${test_result}" || true
   test_status=$?
   if [ $test_status -eq 1 ]; then
     overall_status=1
